Other Losses

Loss functions exist that are not based on distances nor margins. This section lists other useful losses that are implemented in the package:

MisclassLoss

LossFunctions.MisclassLoss — Type

MisclassLoss{R<:AbstractFloat} <: SupervisedLoss

Misclassification loss that assigns 1 for misclassified examples and 0 otherwise. It is a generalization of ZeroOneLoss for more than two classes.

The type parameter R specifies the result type of the loss. Default type is double precision R = Float64.

PoissonLoss

LossFunctions.PoissonLoss — Type

PoissonLoss <: SupervisedLoss

Loss under a Poisson noise distribution (KL-divergence)

$L(output, target) = exp(output) - target*output$

CrossEntropyLoss

LossFunctions.CrossEntropyLoss — Type

CrossEntropyLoss <: SupervisedLoss

The cross-entropy loss is defined as:

$L(output, target) = - target*log(output) - (1-target)*log(1-output)$